+++ formula-android/src/main/java/com/instacart/formula/android/internal/FeatureBinding.kt
@@ -5,7 +5,7 @@ import com.instacart.formula.Formula
 import com.instacart.formula.FormulaContext
 import com.instacart.formula.Stream
 import com.instacart.formula.android.FeatureFactory
-import com.instacart.formula.fragment.FragmentKey
+import com.instacart.formula.android.FragmentKey
 import com.instacart.formula.android.Binding
 import com.instacart.formula.android.FeatureEvent