+++ formula-android/src/main/java/com/instacart/formula/fragment/FragmentEnvironment.kt
@@ -1,5 +1,7 @@
 package com.instacart.formula.fragment
 
+import com.instacart.formula.android.FragmentKey
+
 data class FragmentEnvironment(
     val onScreenError: (FragmentKey, Throwable) -> Unit = { _, it -> throw it }
 )